Refactor: Improve snake_game.py organization
- Move game state management to src/game_state.py
- Move sound management to src/sound.py
- Move UI rendering to src/ui.py
- Simplify main game file

Issue: #1
